requirements.txt
google-api-python-client==2.145.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
tqdm==4.66.5